requirements.txt
streamlit
google.generativeai
python-dotenv
langchain
tiktoken
faiss-cpu
protobuf
langchain_google_genai
InstructorEmbedding

